Added option for coloring encrypted messages
Able to set fish.color.msg_color to any of the available mirc colors, the encrypted messages will then be presented in that color.

This is rather an ugly hack, since weechat has an internal structure for mapping weechat color to mirc color, but currently this is not exposed to the python plugins.